Refactor code from /e2e/helpers
