--- test/e2e/helpers/routeHelpers.ts
@@ -1,24 +1,0 @@
-import { getToken } from '../../helpers/routeHelper.js'
-
-export const get = async (appUrl: string, endpoint: string, headers: Record<string, string> = {}) => {
-  const token = await getToken()
-  const headersWithToken = {
-    authorization: `bearer ${token}`,
-    ...headers,
-  }
-
-  return await fetch(`${appUrl}${endpoint}`, { headers: headersWithToken })
-}
-
-export const del = async (appUrl: string, endpoint: string, headers: Record<string, string> = {}) => {
-  const token = await getToken()
-  const headersWithToken = {
-    authorization: `bearer ${token}`,
-    ...headers,
-  }
-
-  return await fetch(`${appUrl}${endpoint}`, {
-    method: 'DELETE',
-    headers: headersWithToken,
-  })
-}
