+++ test/helpers/registerLogIn.ts
@@ -1,14 +1,14 @@
 import { randomUUID } from 'node:crypto'
 
 import { BrowserContext, expect, Page } from '@playwright/test'
-import { del } from './routeHelpers'
+import { resetUrl } from '../helpers/routeHelper'
 
 export interface CustomBrowserContext extends BrowserContext {
   username?: string
 }
 
 export async function cleanup(urls: string[]) {
-  const results = await Promise.all(urls.map((url) => del(url, '/reset')))
+  const results = await Promise.all(urls.map((url) => resetUrl(url, '/reset')))
   if (!results.every((res) => res.ok)) {
     throw new Error('Error resetting application or deleting emails form smtp4dev')
   }
